@@ -14,15 +14,15 @@
     "start-app-development": "npm start --prefix dashboard"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^2.11.0",
-    "@typescript-eslint/parser": "^2.11.0",
-    "eslint": "6.7.2",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.17.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
+    "@typescript-eslint/eslint-plugin": "~2.34.0",
+    "@typescript-eslint/parser": "~2.34.0",
+    "eslint": "7.0.0",
+    "eslint-config-airbnb": "~18.1.0",
+    "eslint-plugin-import": "~2.20.2",
+    "eslint-plugin-jsx-a11y": "~6.2.3",
+    "eslint-plugin-react": "~7.20.0",
+    "eslint-plugin-react-hooks": "~4.0.2",
     "ini": "1.3.5",
-    "lerna": "3.18.5"
+    "lerna": "~3.21.0"
   }
 }
