@@ -11,57 +11,58 @@
     "styleguide:build": "styleguidist build"
   },
   "resolutions": {
-    "@types/react": "16.9.2"
+    "@types/react": "16.9.35"
   },
   "dependencies": {
-    "@babel/polyfill": "7.6.0",
-    "@material-ui/core": "4.9.0",
-    "@material-ui/icons": "4.4.1",
-    "axios": "0.19.0",
-    "classnames": "2.2.6",
-    "core-js": "3.2.1",
-    "diff": "4.0.1",
-    "dompurify": "2.0.7",
-    "marked": "0.7.0",
-    "moment": "2.24.0",
-    "react-content-loader": "4.2.1",
-    "react-html-parser": "2.0.2",
-    "react-monaco-editor": "0.22.0",
-    "react-router-dom": "4.3.1",
-    "react-truncate-markup": "2.0.1",
-    "typescript": "3.7.3"
+    "@babel/polyfill": "~7.8.7",
+    "@material-ui/core": "~4.9.14",
+    "@material-ui/icons": "~4.9.1",
+    "axios": "~0.19.2",
+    "classnames": "~2.2.6",
+    "core-js": "~3.6.5",
+    "diff": "~4.0.2",
+    "dompurify": "~2.0.11",
+    "marked": "~1.1.0",
+    "moment": "~2.25.3",
+    "react-content-loader": "~5.0.4",
+    "react-html-parser": "~2.0.2",
+    "react-monaco-editor": "~0.36.0",
+    "react-router-dom": "~5.2.0",
+    "react-truncate-markup": "~4.1.1",
+    "typescript": "~3.9.2"
   },
   "devDependencies": {
-    "@babel/core": "7.6.0",
-    "@babel/preset-env": "7.6.0",
-    "@babel/preset-react": "7.0.0",
-    "@types/classnames": "2.2.6",
-    "@types/diff": "3.5.2",
-    "@types/dompurify": "0.0.33",
-    "@types/enzyme": "3.9.4",
-    "@types/jest": "23.3.14",
-    "@types/jsdom": "12.2.1",
-    "@types/marked": "0.6.5",
-    "@types/react": "16.9.2",
-    "@types/react-html-parser": "2.0.1",
-    "@types/react-router-dom": "4.3.1",
-    "@types/react-test-renderer": "16.8.1",
-    "altinn-designsystem": "2.0.19",
-    "babel-jest": "24.7.1",
-    "babel-loader": "8.0.6",
-    "enzyme": "3.10.0",
-    "enzyme-adapter-react-16": "1.14.0",
-    "jest": "23.6.0",
-    "jest-environment-enzyme": "7.0.2",
-    "jest-enzyme": "7.0.2",
-    "jest-junit": "6.3.0",
-    "jsdom": "13.2.0",
-    "react": "16.8.6",
-    "react-docgen-typescript": "1.12.4",
-    "react-dom": "16.8.6",
-    "react-styleguidist": "9.1.2",
-    "react-test-renderer": "16.8.6",
-    "ts-jest": "23.10.5"
+    "@babel/core": "~7.9.6",
+    "@babel/preset-env": "~7.9.6",
+    "@babel/preset-react": "~7.9.4",
+    "@types/classnames": "~2.2.10",
+    "@types/diff": "~4.0.2",
+    "@types/dompurify": "~2.0.2",
+    "@types/enzyme": "~3.10.5",
+    "@types/jest": "~25.2.3",
+    "@types/jsdom": "~16.2.3",
+    "@types/marked": "~0.7.4",
+    "@types/react": "~16.9.35",
+    "@types/react-html-parser": "~2.0.1",
+    "@types/react-router-dom": "~5.1.5",
+    "@types/react-test-renderer": "~16.9.2",
+    "altinn-designsystem": "~2.7.4",
+    "babel-jest": "~26.0.1",
+    "babel-loader": "~8.1.0",
+    "enzyme": "~3.11.0",
+    "enzyme-adapter-react-16": "~1.15.2",
+    "jest": "~26.0.1",
+    "jest-environment-enzyme": "~7.1.2",
+    "jest-enzyme": "~7.1.2",
+    "jest-junit": "~10.0.0",
+    "jsdom": "~16.2.2",
+    "parse5": "6.0.0",
+    "react": "~16.13.1",
+    "react-docgen-typescript": "~1.16.4",
+    "react-dom": "~16.13.1",
+    "react-styleguidist": "~11.0.6",
+    "react-test-renderer": "~16.13.1",
+    "ts-jest": "~26.0.0"
   },
   "jest": {
     "transform": {
