@@ -18,87 +18,82 @@
   "author": "Altinn",
   "license": "3-Clause BSD",
   "resolutions": {
-    "@types/react": "16.9.2"
+    "@types/react": "16.9.35"
   },
   "dependencies": {
-    "@babel/polyfill": "7.6.0",
-    "@material-ui/core": "4.9.0",
-    "@material-ui/icons": "4.4.1",
-    "@types/react-select": "2.0.11",
-    "axios": "0.19.0",
-    "core-js": "3.2.1",
-    "deepmerge": "2.1.1",
-    "dot-object": "1.7.1",
-    "immutability-helper": "2.7.0",
-    "lodash.isequal": "4.5.0",
-    "react": "16.8.6",
-    "react-dnd": "10.0.2",
-    "react-dnd-html5-backend": "10.0.2",
-    "react-dom": "16.8.6",
-    "react-grid-layout": "0.16.6",
-    "react-hot-loader": "4.3.1",
-    "react-modal": "3.4.5",
-    "react-redux": "7.1.0",
-    "react-router-dom": "4.3.1",
-    "react-select": "3.0.8",
-    "redux": "4.0.0",
-    "redux-devtools-extension": "2.13.5",
-    "redux-logger": "3.0.6",
-    "redux-saga": "0.16.0",
-    "reselect": "3.0.1",
-    "typescript": "3.7.3",
-    "uuid": "3.2.1",
-    "webpack": "4.12.0",
-    "webpack-dev-server": "3.1.14"
+    "@babel/polyfill":"~7.8.7",
+    "@material-ui/core":"~4.9.14",
+    "@material-ui/icons":"~4.9.1",
+    "axios":"~0.19.2",
+    "core-js":"~3.6.5",
+    "deepmerge":"~2.1.1",
+    "immutability-helper":"2.7.0",
+    "react":"~16.13.1",
+    "react-dnd":"~10.0.2",
+    "react-dnd-html5-backend":"~10.0.2",
+    "react-dom":"~16.13.1",
+    "react-hot-loader":"~4.12.21",
+    "react-modal":"~3.11.2",
+    "react-redux":"~7.2.0",
+    "react-router-dom":"~5.2.0",
+    "react-select":"~3.1.0",
+    "redux":"~4.0.5",
+    "redux-devtools-extension":"~2.13.8",
+    "redux-logger":"~3.0.6",
+    "redux-saga":"~1.1.3",
+    "reselect":"~4.0.0",
+    "typescript":"~3.9.2",
+    "uuid":"3.2.1"
   },
   "devDependencies": {
-    "@babel/core": "7.6.0",
-    "@babel/preset-env": "7.6.0",
-    "@babel/preset-react": "7.0.0",
-    "@types/classnames": "2.2.6",
-    "@types/deepmerge": "2.1.0",
-    "@types/dot-object": "1.7.0",
-    "@types/enzyme": "3.9.4",
-    "@types/jest": "23.3.14",
-    "@types/lodash.isequal": "4.5.3",
-    "@types/react": "16.9.2",
-    "@types/react-dom": "16.8.3",
-    "@types/react-modal": "3.8.1",
-    "@types/react-redux": "7.1.2",
-    "@types/react-router-dom": "4.3.1",
-    "@types/react-test-renderer": "16.8.1",
-    "@types/redux-mock-store": "1.0.0",
-    "@types/uuid": "3.4.3",
-    "altinn-designsystem": "2.0.19",
-    "awesome-typescript-loader": "5.0.0",
-    "babel-jest": "24.7.1",
-    "classnames": "2.2.6",
-    "cross-env": "5.1.6",
-    "css-loader": "2.1.1",
-    "enzyme": "3.10.0",
-    "enzyme-adapter-react-16": "1.14.0",
-    "eslint": "6.7.2",
-    "eslint-config-airbnb-base": "14.0.0",
-    "html-loader": "0.5.5",
-    "html-webpack-plugin": "3.2.0",
-    "jest": "23.6.0",
-    "jest-environment-enzyme": "7.0.2",
-    "jest-enzyme": "7.0.2",
-    "jest-junit": "6.3.0",
-    "mini-css-extract-plugin": "0.4.0",
-    "prettier": "1.19.1",
-    "prettier-tslint": "0.4.2",
-    "react-test-renderer": "16.8.6",
-    "redux-mock-store": "1.5.3",
-    "resolve-url-loader": "3.0.0",
-    "rimraf": "2.6.2",
-    "source-map-loader": "0.2.3",
-    "style-loader": "0.23.1",
-    "svg-inline-loader": "0.8.0",
-    "ts-jest": "23.10.5",
-    "ts-loader": "4.4.1",
-    "uglifyjs-webpack-plugin": "1.2.5",
-    "webpack-cli": "3.0.3"
+    "@babel/core":"~7.9.6",
+    "@babel/preset-env":"~7.9.6",
+    "@babel/preset-react":"~7.9.4",
+    "@types/classnames":"~2.2.10",
+    "@types/deepmerge":"~2.2.0",
+    "@types/enzyme":"~3.10.5",
+    "@types/jest":"~25.2.3",
+    "@types/react":"16.9.35",
+    "@types/react-dom":"~16.9.8",
+    "@types/react-modal":"~3.10.5",
+    "@types/react-redux":"~7.1.8",
+    "@types/react-router-dom":"~5.1.5",
+    "@types/react-select":"~3.0.13",
+    "@types/react-test-renderer":"~16.9.2",
+    "@types/redux-mock-store":"~1.0.2",
+    "@types/uuid":"3.4.3",
+    "altinn-designsystem":"~2.7.4",
+    "awesome-typescript-loader":"~5.2.1",
+    "babel-jest":"~26.0.1",
+    "classnames":"~2.2.6",
+    "cross-env":"~7.0.2",
+    "css-loader":"~3.5.3",
+    "enzyme":"~3.11.0",
+    "enzyme-adapter-react-16":"~1.15.2",
+    "eslint":"~7.0.0",
+    "eslint-config-airbnb-base":"~14.1.0",
+    "html-loader":"~1.1.0",
+    "html-webpack-plugin":"~4.3.0",
+    "jest":"~26.0.1",
+    "jest-environment-enzyme":"~7.1.2",
+    "jest-enzyme":"~7.1.2",
+    "jest-junit":"~10.0.0",
+    "mini-css-extract-plugin":"~0.9.0",
+    "prettier":"~2.0.5",
+    "prettier-tslint":"~0.4.2",
+    "react-test-renderer":"~16.13.1",
+    "redux-mock-store":"~1.5.4",
+    "resolve-url-loader":"~3.1.1",
+    "rimraf":"~3.0.2",
+    "source-map-loader":"~0.2.4",
+    "style-loader":"~1.2.1",
+    "svg-inline-loader":"~0.8.2",
+    "ts-jest":"~26.0.0",
+    "ts-loader":"~7.0.4",
+    "uglifyjs-webpack-plugin":"~1.2.5",
+    "webpack":"~4.43.0",
+    "webpack-dev-server":"~3.11.0",
+    "webpack-cli":"~3.3.11"
   },
   "jest": {
     "transform": {
