@@ -29,7 +29,7 @@
 
 <div id="root" class="media-body flex-column d-flex"></div>
 <script src="~/designer/js/react/app-development.js" type="text/javascript"></script>
-<script src="~/designer/js/react/0.app-development.js" type="text/javascript"></script>
 <script src="~/designer/js/react/1.app-development.js" type="text/javascript"></script>
 <script src="~/designer/js/react/2.app-development.js" type="text/javascript"></script>
 <script src="~/designer/js/react/3.app-development.js" type="text/javascript"></script>
+<script src="~/designer/js/react/4.app-development.js" type="text/javascript"></script>
