@@ -1,6 +1,6 @@
 // import * as moment from 'moment';
 import { SagaIterator } from 'redux-saga';
-import { delay } from 'redux-saga';
+import { delay } from 'redux-saga/effects';
 import { call, fork, race, take } from 'redux-saga/effects';
 import { get } from 'app-shared/utils/networking';
 import { getAppDeploymentsUrl } from '../../../utils/urlHelper';
