@@ -1,5 +1,6 @@
 import { AxiosError } from 'axios';
-import { delay, SagaIterator } from 'redux-saga';
+import { SagaIterator } from 'redux-saga';
+import { delay } from 'redux-saga/effects';
 import { call, fork, takeLatest } from 'redux-saga/effects';
 import { checkIfAxiosError } from 'app-shared/utils/networking';
 import { post } from '../../../utils/networking';
@@ -25,7 +26,7 @@ function* createReleaseSaga({
     yield call(AppReleaseActionDispatcher.createAppReleaseFulfilled, responseData);
   } catch (err) {
     if (checkIfAxiosError(err)) {
-      const {response: {status}} = err as AxiosError;
+      const { response: { status } } = err as AxiosError;
       yield call(AppReleaseActionDispatcher.createAppReleaseRejected, status);
     }
   }
